refactor: adjust create wallet UI
