+++ src/domains/wallet/components/MnemonicVerification/MnemonicVerification.tsx
@@ -37,7 +37,7 @@ export function MnemonicVerification({ mnemonic, handleComplete }: Properties) {
 	}, [isCompleted]);
 
 	return (
-		<div className="mt-4 grid gap-3 sm:grid-cols-3">
+		<div className="mt-4 grid gap-3 sm:grid-cols-3 w-full">
 			{positions.map((position) => (
 				<MnemonicVerificationInput
 					key={position}
